tests(snapshots): refactored snapshots and added layout snapshot test
#14
