+++ __tests__/snapshots/_components/footer-snapshot.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
-import Alert from '../../src/app/_components/alert'
+import Footer from '../../../src/app/_components/footer'
 
 it('renders homepage unchanged', () => {
-    const { container } = render(<Alert />)
+    const { container } = render(<Footer />)
     expect(container).toMatchSnapshot()
 })
