+++ __tests__/snapshots/_components/more-stories-snapshot.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import { MoreStories } from '../../src/app/_components/more-stories'
+import { MoreStories } from '../../../src/app/_components/more-stories'
 
 
 it('renders homepage unchanged', () => {
