+++ __tests__/snapshots/_components/post-body-snapshot.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import { PostBody } from '../../src/app/_components/post-body'
+import { PostBody } from '../../../src/app/_components/post-body'
 
 it('renders homepage unchanged', () => {
     const { container } = render(<PostBody content={'test'} />)
