+++ __tests__/snapshots/_components/post-title-snapshot.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react'
-import Container from '../../src/app/_components/container'
+import { PostTitle } from '../../../src/app/_components/post-title'
 
 it('renders homepage unchanged', () => {
-    const { container } = render(<Container />)
+    const { container } = render(<PostTitle />)
     expect(container).toMatchSnapshot()
 })
