+++ __tests__/snapshots/_components/section-snapshot.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react'
-import { SectionSeparator } from '../../src/app/_components/section-separator'
+import { SectionSeparator } from '../../../src/app/_components/section-separator'
 
 it('renders homepage unchanged', () => {
     const { container } = render(<SectionSeparator />)
